Fixed auto-reconnect and server-side `close` emission
This commit fixes a bug where clients would attempt to reconnect to the server if the server closed their connection on purpose.
Moreover, this commit fixes a bug where the server's `onClose` event would be fired when a client closes the connection.
Since `onClose` is already reserved for when the `Server` closes, a client `close` can be handled via the `onClientClose` event.