Migrated project to `jest`, added better documentation
The migration to `jest` was pretty horrible. I wouldn't recommend it.

Still, this is the direction that this project had to move into for now, since
it's the only environment for which esm is at least somewhat supported. Even if
the way to support it was going through a terrible hack that can be found in
`./src/esm.mjs`. As soon as the esm support is more mature, I can definitely
see myself porting the project back to `ava`.

I also added some better documentation to this commit. The server events are
still largely undocumented and will follow at a later point.